+++ connet/src/test/java/houseInception/connet/service/PrivateRoomServiceTest.java
@@ -5,10 +5,9 @@ import houseInception.connet.domain.privateRoom.PrivateChat;
 import houseInception.connet.domain.privateRoom.PrivateRoom;
 import houseInception.connet.domain.privateRoom.PrivateRoomUser;
 import houseInception.connet.dto.*;
+import houseInception.connet.dto.chatEmoji.ChatEmojiResDto;
 import houseInception.connet.exception.PrivateRoomException;
 import houseInception.connet.repository.PrivateRoomRepository;
-import houseInception.connet.repository.UserBlockRepository;
-import houseInception.connet.repository.UserRepository;
 import jakarta.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
